fix(options): `.with` breaks with menu params
Previously, a menu param, if set to a value not included in the options,
results in an erroneous item unable to parse (thus disappars from the
param list).

Now `.with` won't set the param if the value is neither one of the
option values nor an 0-based index of an option.